refactor/create app.go
