+++ cart/Makefile
@@ -1,0 +1,5 @@
+@build: 
+	docker-compose build
+
+@dev: 
+	docker-compose up -d
